+++ cooked_input/__init__.py
@@ -1,11 +1,13 @@
 # -*- coding: utf-8 -*-
 
-__version__ = '0.2.0'
+__version__ = '0.2.1'
 
 from .get_input import get_input, get_table_input, process_value, make_pretty_table
-from .get_input import TABLE_ID, TABLE_VALUE, TABLE_ID_OR_VALUE
+from .get_input import get_string, get_int, get_float, get_boolean, get_date, get_yes_no, get_list
+from .error_callbacks import MaxRetriesExceeded
 from .error_callbacks import print_error, log_error, silent_error, DEFAULT_CONVERTOR_ERROR, DEFAULT_VALIDATOR_ERROR
 
+from .convertors import TABLE_ID, TABLE_VALUE, TABLE_ID_OR_VALUE
 from .convertors import Convertor, IntConvertor, FloatConvertor, BooleanConvertor
 from .convertors import ListConvertor, DateConvertor, YesNoConvertor, TableConvertor
 from .validators import Validator, ExactLengthValidator, InLengthValidator, ExactValueValidator, InRangeValidator